refactor(framework): useTransaction/createTransaction→usePlugin/createPlugin, flatten options, drop subRuns (solid/vue/svelte)
