+++ packages/ai-solid/tests/use-plugin.test.tsx
@@ -1,17 +1,17 @@
 import { renderHook, waitFor } from '@solidjs/testing-library'
 import { describe, expect, expectTypeOf, it } from 'vitest'
-import { chatVerb, defineTransaction, verb } from '@tanstack/ai/transaction'
+import { chatPlugin, definePlugin, generationPlugin } from '@tanstack/ai/plugin'
 import { stream } from '@tanstack/ai-client'
 import { z } from 'zod'
-import { useTransaction } from '../src/use-transaction.js'
+import { usePlugin } from '../src/use-plugin.js'
 
-// A connection adapter that replays canned chunks per verb, branching on the
-// `verb` discriminator forwarded by TransactionClient.
+// A connection adapter that replays canned chunks per plugin, branching on the
+// `plugin` discriminator forwarded by PluginClient.
 function fakeConnection() {
   return stream(async function* (_messages, data) {
-    const verbName = (data as Record<string, unknown> | undefined)?.verb
+    const pluginName = (data as Record<string, unknown> | undefined)?.plugin
 
-    if (verbName === 'primaryChat') {
+    if (pluginName === 'primaryChat') {
       yield { type: 'RUN_STARTED', threadId: 't', runId: 'r' } as any
       yield {
         type: 'TEXT_MESSAGE_START',
@@ -25,29 +25,37 @@ function fakeConnection() {
       } as any
       yield { type: 'TEXT_MESSAGE_END', messageId: 'm1' } as any
       yield { type: 'RUN_FINISHED', threadId: 't', runId: 'r' } as any
-    } else if (verbName === 'blogPost') {
-      // A transaction run: one banner sub-run, then the final result.
+    } else {
       yield { type: 'RUN_STARTED', threadId: 't', runId: 'r' } as any
       yield {
         type: 'CUSTOM',
-        name: 'transaction:sub-run:started',
-        value: { runId: 'r-sub-0', parentRunId: 'r', verb: 'banner', index: 0 },
-      } as any
-      yield {
-        type: 'CUSTOM',
-        name: 'transaction:sub-run:result',
-        value: {
-          runId: 'r-sub-0',
-          verb: 'banner',
-          index: 0,
-          result: { url: 'hero.png' },
-        },
-      } as any
-      yield {
-        type: 'CUSTOM',
         name: 'generation:result',
-        value: { title: 'Foxes', heroUrl: 'hero.png' },
+        value: { url: 'u' },
       } as any
+      yield { type: 'RUN_FINISHED', threadId: 't', runId: 'r' } as any
+    }
+  })
+}
+
+// A connection adapter that records every request body it sees, so tests can
+// assert per-plugin options (forwardedProps) reached the client's request.
+function capturingConnection(seen: Array<Record<string, unknown>>) {
+  return stream(async function* (_messages, data) {
+    seen.push((data ?? {}) as Record<string, unknown>)
+    const pluginName = (data as Record<string, unknown> | undefined)?.plugin
+    if (pluginName === 'primaryChat') {
+      yield { type: 'RUN_STARTED', threadId: 't', runId: 'r' } as any
+      yield {
+        type: 'TEXT_MESSAGE_START',
+        messageId: 'm1',
+        role: 'assistant',
+      } as any
+      yield {
+        type: 'TEXT_MESSAGE_CONTENT',
+        messageId: 'm1',
+        delta: 'hello',
+      } as any
+      yield { type: 'TEXT_MESSAGE_END', messageId: 'm1' } as any
       yield { type: 'RUN_FINISHED', threadId: 't', runId: 'r' } as any
     } else {
       yield { type: 'RUN_STARTED', threadId: 't', runId: 'r' } as any
@@ -61,48 +69,39 @@ function fakeConnection() {
   })
 }
 
-function makeTransaction() {
-  const banner = verb({
-    input: z.object({ prompt: z.string() }),
-    execute: async ({ input }) => ({ url: `img:${input.prompt}` }),
-  })
-  return defineTransaction({
-    primaryChat: chatVerb(async function* (r: any) {
+function makePlugin() {
+  return definePlugin({
+    primaryChat: chatPlugin(async function* (r: any) {
       yield { type: 'RUN_STARTED', threadId: r.threadId, runId: r.runId } as any
     } as any),
-    banner,
-    blogPost: verb({
-      input: z.object({ topic: z.string() }),
-      execute: async ({ input }, ctx) => {
-        const hero = await ctx.call(banner, { prompt: input.topic })
-        return { title: input.topic, heroUrl: hero.url }
-      },
+    banner: generationPlugin({
+      input: z.object({ prompt: z.string() }),
+      execute: async ({ input }) => ({ url: `img:${input.prompt}` }),
     }),
   })
 }
 
-describe('useTransaction', () => {
-  it('exposes only the declared verbs, by kind', () => {
-    const txn = makeTransaction()
+describe('usePlugin', () => {
+  it('exposes only the declared plugins, by kind', () => {
+    const def = makePlugin()
     const { result } = renderHook(() =>
-      useTransaction(txn, { connection: fakeConnection() }),
+      usePlugin(def, { connection: fakeConnection() }),
     )
 
     expect(result.primaryChat).toBeDefined()
     expect(result.primaryChat.sendMessage).toBeDefined()
     expect(result.banner).toBeDefined()
     expect(result.banner.run).toBeDefined()
-    expect(result.blogPost).toBeDefined()
     expect((result as any).speech).toBeUndefined()
   })
 
-  it('run() on a one-shot verb populates its result, typed by the schema', async () => {
-    const txn = makeTransaction()
+  it('run() on a generation plugin populates its result, typed by the schema', async () => {
+    const def = makePlugin()
     const { result } = renderHook(() =>
-      useTransaction(txn, { connection: fakeConnection() }),
+      usePlugin(def, { connection: fakeConnection() }),
     )
 
-    // The input type comes from the verb's schema.
+    // The input type comes from the plugin's schema.
     expectTypeOf(result.banner.run)
       .parameter(0)
       .toEqualTypeOf<{ prompt: string }>()
@@ -121,10 +120,10 @@ describe('useTransaction', () => {
     expect(result.banner.isLoading).toBe(false)
   })
 
-  it('sendMessage() on a chat verb populates its messages', async () => {
-    const txn = makeTransaction()
+  it('sendMessage() on a chat plugin populates its messages', async () => {
+    const def = makePlugin()
     const { result } = renderHook(() =>
-      useTransaction(txn, { connection: fakeConnection() }),
+      usePlugin(def, { connection: fakeConnection() }),
     )
 
     expect(result.primaryChat.messages).toEqual([])
@@ -154,51 +153,29 @@ describe('useTransaction', () => {
     expect(textPart?.content).toBe('hello')
   })
 
-  it('surfaces live sub-run state on a transaction run', async () => {
-    const txn = makeTransaction()
-    const { result } = renderHook(() =>
-      useTransaction(txn, { connection: fakeConnection() }),
-    )
-
-    await result.blogPost.run({ topic: 'foxes' })
-
-    expect(result.blogPost.result).toEqual({
-      title: 'Foxes',
-      heroUrl: 'hero.png',
-    })
-    expect(result.blogPost.subRuns).toHaveLength(1)
-    expect(result.blogPost.subRuns[0]).toMatchObject({
-      verb: 'banner',
-      status: 'success',
-      result: { url: 'hero.png' },
-    })
-    // The sibling banner surface is untouched by the transaction's sub-run.
-    expect(result.banner.result).toBeNull()
-  })
-
   it('exposes chat partial/final with cleared defaults on first render', () => {
-    const txn = makeTransaction()
+    const def = makePlugin()
     const { result } = renderHook(() =>
-      useTransaction(txn, { connection: fakeConnection() }),
+      usePlugin(def, { connection: fakeConnection() }),
     )
 
     expect((result.primaryChat as any).partial).toEqual({})
     expect((result.primaryChat as any).final).toBeNull()
   })
 
-  it('applies a one-shot onResult transform and infers its type', async () => {
-    const txn = makeTransaction()
+  it('applies a generation onResult transform and infers its type', async () => {
+    const def = makePlugin()
     const { result } = renderHook(() =>
-      useTransaction(txn, {
+      // Flat options: the per-plugin `banner` entry sits at the top level
+      // alongside `connection`, no nested `plugins`/`verbs` wrapper.
+      usePlugin(def, {
         connection: fakeConnection(),
-        verbs: {
-          banner: {
-            onResult: (raw) => {
-              // The transform's input is the verb's typed result
-              // (`toEqualTypeOf` fails if `raw` were `any`).
-              expectTypeOf(raw).toEqualTypeOf<{ url: string }>()
-              return raw.url
-            },
+        banner: {
+          onResult: (raw) => {
+            // The transform's input is the plugin's typed result
+            // (`toEqualTypeOf` fails if `raw` were `any`).
+            expectTypeOf(raw).toEqualTypeOf<{ url: string }>()
+            return raw.url
           },
         },
       }),
@@ -212,10 +189,35 @@ describe('useTransaction', () => {
     expect(result.banner.result).toBe('u')
   })
 
+  it('threads flat per-plugin options through to the client', async () => {
+    const def = makePlugin()
+    const seen: Array<Record<string, unknown>> = []
+    const { result } = renderHook(() =>
+      // A chat plugin's `forwardedProps` AND a generation plugin's `onResult`
+      // provided as sibling top-level keys — proving the flat shape both wires
+      // the request body and threads the result transform.
+      usePlugin(def, {
+        connection: capturingConnection(seen),
+        primaryChat: { forwardedProps: { tone: 'punchy' } },
+        banner: { onResult: (raw) => `url:${raw.url}` },
+      }),
+    )
+
+    await result.primaryChat.sendMessage('hi')
+    await result.banner.run({ prompt: 'a fox' })
+
+    // The chat plugin's forwardedProps reached the request body alongside the
+    // routing discriminator.
+    const chatData = seen.find((d) => d.plugin === 'primaryChat')
+    expect(chatData?.tone).toBe('punchy')
+    // The generation plugin's onResult transform was applied to its result.
+    expect(result.banner.result).toBe('url:u')
+  })
+
   it('disposes every chat and one-shot sub-client on cleanup', () => {
-    const txn = makeTransaction()
+    const def = makePlugin()
     const { result, cleanup } = renderHook(() =>
-      useTransaction(txn, { connection: fakeConnection() }),
+      usePlugin(def, { connection: fakeConnection() }),
     )
 
     expect(result.primaryChat).toBeDefined()
