+++ packages/ai-svelte/src/plugin.ts
@@ -1,0 +1,1 @@
+export { createPlugin } from './create-plugin.svelte'
