--- packages/ai-svelte/src/transaction.ts
@@ -1,1 +1,0 @@
-export { createTransaction } from './create-transaction.svelte'
