+++ packages/ai-vue/src/plugin.ts
@@ -1,0 +1,1 @@
+export { usePlugin } from './use-plugin'
