--- packages/ai-vue/src/transaction.ts
@@ -1,1 +1,0 @@
-export { useTransaction } from './use-transaction'
