+++ packages/ai-vue/tsdown.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from 'tsdown'
 
 export default defineConfig({
-  entry: ['./src/index.ts', './src/transaction.ts'],
+  entry: ['./src/index.ts', './src/plugin.ts'],
   format: ['esm'],
   unbundle: true,
   dts: true,
